% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_PCE_TDS100.R
\name{.getMetadata}
\alias{.getMetadata}
\title{Get Meta Data from File Lines}
\usage{
.getMetadata(
  mylines,
  beginPattern = "Outer Diameter",
  endPattern = "Txducers Spacing",
  transpose = TRUE
)
}
\arguments{
\item{mylines}{vector of character, each representing a line of a text file}

\item{beginPattern}{pattern matching the line in which meta data block begins}

\item{endPattern}{pattern matching the line in which meta data block ends}

\item{transpose}{(default: TRUE)}
}
\value{
meta data read from \emph{mylines} in the form of a 
  data frame with one row containing the meta data values in the different
  columns (transpose == FALSE) or a data frame with columns \emph{names} 
  and \emph{values} (transpose == TRUE)
}
\description{
Lines between lines containing "Outer Diameter" and  "Txducers Spacing", 
respectively, are extrated.
}
