% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_Ori_MLog.R
\name{.getSampleTimes}
\alias{.getSampleTimes}
\title{Read Sample Times from Automatic Sampler File}
\usage{
.getSampleTimes(
  filepath,
  blockbegin,
  sep = ";",
  dec = ",",
  captionPattern = "Datum;Uhrzeit",
  warn = TRUE
)
}
\arguments{
\item{filepath}{full path to file generated by automatic sample}

\item{blockbegin}{blockbegin}

\item{sep}{field separator (default: ";")}

\item{dec}{decimal separator (default: ",")}

\item{captionPattern}{pattern matching the caption line. Default:
"Datum;Uhrzeit"}

\item{warn}{should warnings be printed (default: TRUE)}
}
\description{
Read Sample Times from Automatic Sampler File
}
