% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_PCE_TDS100.R
\name{.getValidTimes}
\alias{.getValidTimes}
\title{Read Times from Given Lines}
\usage{
.getValidTimes(mylines, timeformat = "\%y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{mylines}{vector of character, each representing a line of a text file}

\item{timeformat}{time format to be looked for in \emph{mylines}. Default:
"\%y-\%m-\%d \%H:\%M:\%S"}
}
\value{
list with elements \emph{tstamps} (timestamps as text in ISO format)
  and \emph{tstampIndex} (contains for each text line the index in 
  \emph{tstamps} at which the timestamp corresponding to the line is found.
}
\description{
Read Times from Given Lines
}
