% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_General.R
\name{extdataFile}
\alias{extdataFile}
\title{Path to File in Installed Package}
\usage{
extdataFile(..., must_exist = TRUE, dbg = FALSE)
}
\arguments{
\item{...}{parts of the file path to be passed to \code{\link{system.file}}}

\item{must_exist}{if \code{TRUE} (the default) and the specified file does 
not exist, the program stops with an error message}

\item{dbg}{if \code{TRUE} (the default is \code{FALSE}) debug messages are 
shown}
}
\description{
Path to File in Installed Package
}
\examples{
# List the files provided in the "extdata" folder of kwb.logger
dir(extdataFile())
}
