% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_timestamps.R
\name{getExampleTimestamps}
\alias{getExampleTimestamps}
\title{Timestamp Strings from Example Looger Files}
\usage{
getExampleTimestamps()
}
\description{
Timestamp Strings from Example Looger Files
}
