% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_Aquatroll.R
\name{read_aquatroll_data}
\alias{read_aquatroll_data}
\title{Read Aquatroll Data}
\usage{
read_aquatroll_data(file)
}
\arguments{
\item{file}{full path to logger file}
}
\description{
Use readLogger_InSituInc_Aquatroll instead!
}
