% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_LT_EDGE_M100.R
\name{read_logger_LT_EDGE_M100}
\alias{read_logger_LT_EDGE_M100}
\title{Read Logger File from LT EDGE M100}
\usage{
read_logger_LT_EDGE_M100(
  file,
  timeFormat = .defaultTimeFormat("v4"),
  metaToColumns = FALSE,
  dbg = TRUE,
  ...
)
}
\arguments{
\item{file}{full path to the file (either .xle or .csv)}

\item{timeFormat}{time format in which the timestamps are given in the file.
Default: "yyyy/mm/dd HH:MM:SS". It is assumed that the timestamps represent
Berlin normal time = Berlin winter time (no daylight saving)}

\item{metaToColumns}{if TRUE, metadata are written into the last columns of
the returned data frame}

\item{dbg}{if TRUE (default) debug messages (e.g. "Reading...") are shown}

\item{\dots}{arguments passed to \code{\link{read_logger_LT_EDGE_M100_csv}}
or \code{\link{read_logger_LT_EDGE_M100_xle}}}
}
\value{
data frame with columns \emph{TimestampInFile} (character timestamp
  as it appeared in the file), \emph{DateTimeUTC} (POSIXct timestamp in time
  zone "UTC"), \emph{LocalDateTime} (POSIXct timestamp in time zone 
  "Europe/Berlin"), \emph{UTCOffset} (integer Offset in hours between UTC and
  local time), \emph{Level} (water level in unit as given in the file 
  header), \emph{Temperature} (water temperature in unit as given in the file
  header)
}
\description{
Read an input file of water level logger "LT EDGE M100"
}
