% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_LT_EDGE_M100.R
\name{read_logger_LT_EDGE_M100_csv}
\alias{read_logger_LT_EDGE_M100_csv}
\title{Read Logger File (.csv) from LT EDGE M100}
\usage{
read_logger_LT_EDGE_M100_csv(file, country = "en", maxRowToLookForHeader = 15)
}
\arguments{
\item{file}{full path to logger file}

\item{country}{one of "en" (English format: column separator = ",", decimal
sign = ".") or "de" (German format: column separator = ";", decimal sign =
",")}

\item{maxRowToLookForHeader}{number of rows to look for column headers
(default: 15)}
}
\description{
Read an input file of water level logger "LT EDGE M100" from .csv file
}
