% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_Buehler_4010.R
\name{readLogger_Buehler_4010_Flaschenbericht}
\alias{readLogger_Buehler_4010_Flaschenbericht}
\title{Read Logger File from Buehler 4010 (Flaschenbericht)}
\usage{
readLogger_Buehler_4010_Flaschenbericht(
  samplerFile,
  headerPattern = paste0("Flasche\\\\s+Datum\\\\s+Zeit\\\\s+F.llzeit\\\\s+PN Soll",
    "\\\\s+PN Ist\\\\s+P/E"),
  endPattern = "---",
  tz = "Etc/GMT+1"
)
}
\arguments{
\item{samplerFile}{full path to file generated by auto sampler}

\item{headerPattern}{pattern matching the header row}

\item{endPattern}{pattern matching the end of the table}

\item{tz}{time zone. Default: "Etc/GMT+1}
}
\description{
Read Logger File from Buehler 4010 (Flaschenbericht)
}
