% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_Buehler_4010.R
\name{readLogger_Buehler_4010_Infospeicher}
\alias{readLogger_Buehler_4010_Infospeicher}
\title{Read Logger File from Buehler 4010 (Infospeicher)}
\usage{
readLogger_Buehler_4010_Infospeicher(samplerFile, unprocessed = TRUE, ...)
}
\arguments{
\item{samplerFile}{full path to file generated by auto sampler Buehler 4010
(Infospeicher)}

\item{unprocessed}{if TRUE, the information is returned "raw", i.e. without
any aggregation}

\item{\dots}{arguments passed to \code{.read_Infospeicher.raw}, e.g.
\emph{tz}, see there}
}
\description{
Read Logger File from Buehler 4010 (Infospeicher)
}
