% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_FLEXIM_F601.R
\name{readLogger_FLEXIM_F601}
\alias{readLogger_FLEXIM_F601}
\title{Read Logger File from FLEXIM F601}
\usage{
readLogger_FLEXIM_F601(
  filename,
  sep = "\\t",
  dec = ",",
  timeformat = "\%d.\%m.\%Y \%H:\%M:\%S",
  headerPattern = "Uhrzeit\\tDruck",
  date_yyyymmdd = substr(basename(filename), 1, 8),
  dbg = TRUE
)
}
\arguments{
\item{filename}{full path to logger file}

\item{sep}{column separator}

\item{dec}{decimal character}

\item{timeformat}{time format string}

\item{headerPattern}{pattern matching the header line}

\item{date_yyyymmdd}{datestring, by default taken from the filename}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Read Logger File from FLEXIM F601
}
\examples{
# Set path to example file (contained in this package)
file <- extdataFile("FLEXIM/example_FLEXIM_F601_short.txt")
   
# Let's have a look on the file structure
writeLines(readLines(file))

\dontrun{   
# Now read the file
x <- readLogger_FLEXIM_F601(file)
  
# Show the first lines
head(x)
  
# Get the meta data and show its structure
str(kwb.utils::getAttribute(x, "metadata"))
}
}
\references{
\url{http://www.flexim.com/files/tsfluxus_f601v1-5-1de_leu.pdf}
}
