% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_others.R
\name{readLogger_FLUKE_1730}
\alias{readLogger_FLUKE_1730}
\title{Read Logger File from FLUKE_1730}
\usage{
readLogger_FLUKE_1730(filepath, sep = ";", dec = ".")
}
\arguments{
\item{filepath}{full path to logger file}

\item{sep}{column separator}

\item{dec}{decimal character}
}
\description{
Read Logger File from FLUKE_1730
}
\examples{
\dontrun{
# set path to example file (contained in this package)
file <- extdataFile("FLUKE/example_FLUKE_1730_1.txt")

# read the file
x <- readLogger_FLUKE_1730(file)
  
# examine the list structure of the result
str(x)  
}
}
\references{
\url{http://assets.fluke.com/manuals/1730____umeng0000.pdf}
}
