% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_Aquatroll.R
\name{readLogger_InSituInc_Aquatroll}
\alias{readLogger_InSituInc_Aquatroll}
\title{Read File From InSituInc Aquatroll}
\usage{
readLogger_InSituInc_Aquatroll(
  csv,
  headerPattern = NULL,
  timestampFormat = c("\%d.\%m.\%Y \%H:\%M:\%S", "\%d.\%m.\%Y\%H:\%M:\%S"),
  tz = "Etc/GMT+1",
  maxRowToLookForHeader = 700,
  model = "",
  dbg = FALSE,
  fileEncoding = "latin1"
)
}
\arguments{
\item{csv}{full path to logger file}

\item{headerPattern}{pattern matching the header of the table. See source
code for the default.}

\item{timestampFormat}{vector of possible timestamp formats. See help for the
default.}

\item{tz}{time zone. default: "Etc/GMT+1"}

\item{maxRowToLookForHeader}{number of first rows in the file to read in
advance to look for the header line. Default: 700}

\item{model}{model of the Aquatroll (either "" or "600")}

\item{dbg}{if \code{TRUE}, debug messages are shown}

\item{fileEncoding}{encoding of the input file. Default: "latin1"}
}
\value{
if model = "600" data frame with columns \code{DateTime}, 
  \code{Seconds}, \code{Temp}, \code{ActCond}, \code{SpecCond},
  \code{TDS}, \code{Resis}, \code{WaterD}, \code{DOC}, \code{DOP},
  \code{Tur}, \code{Sal} otherwise data frame with columns
  \code{DateTime}, \code{WaterTemp.C}, \code{SpecCond.uS}
}
\description{
Read File From InSituInc Aquatroll
}
\examples{
# set path to example file (contained in this package)
file_1 <- extdataFile("InSituInc/example_InSituInc_Aquatroll.csv")
file_2 <- extdataFile("InSituInc/example_InSituInc_Aquatroll_600.csv")

# let's have a look on the file structure (read first 75 lines)
writeLines(kwb.utils::readLinesWithEncoding(file_1, 75, fileEncoding = "latin1"))
  
# now read the file
x <- readLogger_InSituInc_Aquatroll(csv = file_1)
  
# show the first lines
head(x)
  
# Read a file of Aquatroll, model "600"
x600 <- readLogger_InSituInc_Aquatroll(csv = file_2, model = "600")
  
# show the first lines
head(x600)
  
# show the structure
str(x600)

}
