% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_NIVUS_PCM4.R
\name{readLogger_NIVUS_PCM4}
\alias{readLogger_NIVUS_PCM4}
\title{Read Logger File from NIVUS PCM4}
\usage{
readLogger_NIVUS_PCM4(csv, completenessRequiredFor = c("DateTime", "H", "v"))
}
\arguments{
\item{csv}{full path to CSV file}

\item{completenessRequiredFor}{character string of column names that are
required not to be empty}
}
\description{
Read Logger File from NIVUS PCM4
}
\examples{
# set path to example file (contained in this package)
filepath <- grep("NIVUS_PCM4", exampleLoggerFiles(), value = TRUE)[1]
  
# read the file
x <- readLogger_NIVUS_PCM4(filepath)
  
# examine the list structure of the result
str(x)  

}
\references{
\url{http://www.nivus.de/ximages/1397007_p4ba02en.pdf}
}
