% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_Ori_MLog.R
\name{readLogger_Ori_BasicEx1}
\alias{readLogger_Ori_BasicEx1}
\title{Read Logger File from Ori BasicEx1}
\usage{
readLogger_Ori_BasicEx1(
  filepath,
  infotype = validInfoTypes(),
  blockbegin = "ORI BasicEx1",
  warn = TRUE,
  sep = ";",
  dec = ",",
  colnameDate = "Datum",
  colnameTime = "Uhrzeit",
  dateformat = .defaultTimeFormat("v5"),
  timeformat = .defaultTimeFormat("v1")
)
}
\arguments{
\item{filepath}{full path to file generated by automatic sampler}

\item{infotype}{one or more of the values returned by \code{validInfoTypes}}

\item{blockbegin}{identification of "block begins"; Default: "ORI BasicEx1"}

\item{warn}{if TRUE, warnings are generated if a block does not contain
"Probe"}

\item{sep}{column separator}

\item{dec}{decimal character}

\item{colnameDate}{name of date column}

\item{colnameTime}{name of time column}

\item{dateformat}{date format string}

\item{timeformat}{time format string}
}
\description{
Read Logger File from Ori BasicEx1
}
\examples{
\dontrun{
# set path to example file (contained in this package)
(file <- extdataFile("Ori/example_Ori_BasicEx1.csv"))

# read the "actions" from the file
readLogger_Ori_BasicEx1(file, infotype = "actions")
  
# read the sample times from the file
readLogger_Ori_BasicEx1(
  file, infotype = "times", blockbegin = "ORI BasicEx1 TU Berlin"
)
  
# read both at the same time
x <- readLogger_Ori_BasicEx1(
  file, blockbegin = "ORI BasicEx1 TU Berlin"
)
  
# examine the list structure of the result
str(x)  
}

}
\references{
\url{http://www.origmbh.de/fileadmin/user_upload/pdf/basic_ex_1_mobil/ORI_Basic_Ex1_mobil_de.pdf}
}
