% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_Ori_MLog.R
\name{readLogger_Ori_MLog}
\alias{readLogger_Ori_MLog}
\title{Read Water Level File from Radar Probe}
\usage{
readLogger_Ori_MLog(csv, sep = "\\t", dec = ",",
  timeFormat = c(.defaultTimeFormat("v8"), .defaultTimeFormat("v7")),
  ...)
}
\arguments{
\item{csv}{full path to file generated by radar probe}

\item{sep}{column separator. Default: tabulator character}

\item{dec}{decimal character. Default: ","}

\item{timeFormat}{time format string. Default: "\%d.\%m.\%Y \%H:\%M:\%S"}

\item{\dots}{further arguments passed to \code{readCsvInputFile}, e.g. 
\emph{stopOnMissingColumns}}
}
\description{
Read Water Level File from Radar Probe
}
\examples{
  filepath <- grep("Ori_MLog", exampleLoggerFiles(), value=TRUE)[1]
  
  # set separator different from default (tabulator) and do not complain
  # about missing columns
  x <- readLogger_Ori_MLog(filepath, sep=";", stopOnMissingColumns=FALSE)
  
  str(x)
  
}
\references{
\url{http://www.origmbh.de/fileadmin/user_upload/produkte/sensoren/DA_MLOG_Vega_Radar_Puls_61_en.pdf}
}
