% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_others.R
\name{readLogger_PCE_PA8000}
\alias{readLogger_PCE_PA8000}
\title{Read Logger File from PCE PA8000}
\usage{
readLogger_PCE_PA8000(
  filename,
  sep = "\\t",
  dec = ",",
  timeformat = .defaultTimeFormat("v2"),
  headerPattern = "Position\\tDate\\tTime"
)
}
\arguments{
\item{filename}{full path to logger file}

\item{sep}{column separator}

\item{dec}{decimal character}

\item{timeformat}{time format string}

\item{headerPattern}{pattern matching the table header row}
}
\description{
Read Logger File from PCE PA8000
}
\examples{
\dontrun{
# set path to example file (contained in this package)
(file <- extdataFile("PCE/example_PCE_PA8000.txt"))

# read the file
x <- readLogger_PCE_PA8000(file)
  
# examine the structure of the result
str(x)  
}

}
\references{
\url{http://www.industrial-needs.com/manual/manual-pce-pa8000.pdf}
}
