% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_PCE_TDS100.R
\name{readLogger_PCE_TDS100}
\alias{readLogger_PCE_TDS100}
\title{Read Logger File from PCE TDS100}
\usage{
readLogger_PCE_TDS100(txt, timeformat = .defaultTimeFormat("v3"))
}
\arguments{
\item{txt}{full path to text file generated by logger}

\item{timeformat}{Date and time format used in the file. 
Default: "\%y-\%m-\%d \%H:\%M:\%S"}
}
\value{
data frame with columns \emph{tstamp}, \emph{flow}, \emph{flowunit}, 
  \emph{vel}, \emph{velunit}, \emph{UP}, \emph{DN}, \emph{Q}. Duplicate lines
  are removed. The data frame has an attribute \emph{metadata} containing the
  meta data of the file, as returned by \code{.getMetadata}
}
\description{
Read logger file of flow meter TDS 100 by PCE Deutschland GmbH
}
\examples{
\dontrun{
# set path to example file (contained in this package)
file_1 <- extdataFile("PCE/example_PCE_TDS100.log")
file_2 <- extdataFile("PCE/example_PCE_TDS100_noMeta.log")

# read a file containing metadata
x1 <- readLogger_PCE_TDS100(file_1) # warnings about duplicate timestamps
  
# read a file not containing metadata
x2 <- readLogger_PCE_TDS100(file_2) # warning about missing meta data
  
# examine the structures of the results
str(x1)
str(x2)
  
# get meta data from attribute "metadata"
kwb.utils::getAttribute(x1, "metadata")
}

}
\references{
\url{http://www.warensortiment.de/bedienung/ba_durchflussmessgeraet-pce-tds100h_de-v-1-1.pdf}

\url{http://www.industrial-needs.com/manual/manual-pce-tds-100h.pdf}
}
