% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_SIGMA.R
\name{readLogger_SIGMA_SD900}
\alias{readLogger_SIGMA_SD900}
\title{Read Logger File from SIGMA_SD900}
\usage{
readLogger_SIGMA_SD900(filepath, successOnly = FALSE, sep = ",",
  dateformat = .defaultTimeFormat("v6"))
}
\arguments{
\item{filepath}{full path to input file}

\item{successOnly}{if TRUE, only file entries with "SUCCESS" in column
"-RESULT-" are returned. Default: FALSE}

\item{sep}{column separator. Default: ","}

\item{dateformat}{format of timestamp. Default: "\%H:\%M \%m/\%d/\%Y"}
}
\description{
Read Logger File from SIGMA_SD900
}
\examples{
# set path to example file (contained in this package)
(filepath <- grep("SIGMA_SD900", exampleLoggerFiles(), value = TRUE))
  
# read the file
(samples <- readLogger_SIGMA_SD900(filepath))
  
# read only lines representing successful samples
(samplesOk <- readLogger_SIGMA_SD900(filepath, successOnly = TRUE))
  
# show metadata (given in attribute "metadata")
kwb.utils::getAttribute(samplesOk, "metadata")    

}
\references{
\url{http://www.hach.com/asset-get.download.jsa?id=7639983273}
}
