% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_VEGABAR.R
\name{readLogger_VEGA_BAR54_raw}
\alias{readLogger_VEGA_BAR54_raw}
\title{Read Logger File from VEGA BAR54 raw}
\usage{
readLogger_VEGA_BAR54_raw(filepath, trimMetadata = TRUE)
}
\arguments{
\item{filepath}{full path to logger file}

\item{trimMetadata}{if TRUE, metadata of class "character" are trimmed (by
using \code{\link[kwb.utils]{hsTrim}})}
}
\description{
Read Logger File from VEGA BAR54 raw
}
\examples{
# set paths to example files (contained in this package)
(filepaths <- grep("VEGA_BAR54_raw", exampleLoggerFiles(), value = TRUE))
  
# read the files
x1 <- readLogger_VEGA_BAR54_raw(filepaths[1])
x2 <- readLogger_VEGA_BAR54_raw(filepaths[2])
    
# get meta data
kwb.utils::getAttribute(x1, "metadata")
kwb.utils::getAttribute(x2, "metadata")

}
\seealso{
\code{\link{readLogger_VEGA_BAR54}}
}
