% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger_VEGABAR.R
\name{readLogger_VEGA_BAR54}
\alias{readLogger_VEGA_BAR54}
\title{Read Logger File from VEGA BAR54}
\usage{
readLogger_VEGA_BAR54(filename, date_yyyymmdd = "", sep = "\\t",
  dec = ",", timeformat = .defaultTimeFormat("v2"),
  headerPattern = "Uhrzeit\\tDruck")
}
\arguments{
\item{filename}{full path to inoput file}

\item{date_yyyymmdd}{day of measurements as text in format "yyyymmdd", e.g.
"20140423" for April 23 of 2014}

\item{sep}{column separator}

\item{dec}{decimal character}

\item{timeformat}{time format string}

\item{headerPattern}{pattern matching the table header row}
}
\description{
Read Logger File from VEGA BAR54
}
\examples{
# set path to example file (contained in this package)
(filepath <- grep("VEGA_BAR54[.]", exampleLoggerFiles(), value = TRUE))
  
# read the file
x <- readLogger_VEGA_BAR54(filepath)
  
# examine the structure of the result
str(x)  

}
\references{
\url{http://www.vega.com/downloads/PI/EN/37528-EN.PDF}
}
\seealso{
\code{\link{readLogger_VEGA_BAR54_raw}}
}
